Sending RGB instead of RGBA images
We need to send and read 3/4 bytes compared to before. Things should be
faster. Transitions should also be faster now.
