+++ TODO
@@ -1,5 +1,3 @@
-Update sctk
-
 Fully automated, complete testing, for every option
 
-We are busy waiting with thread sleep. That's an actual problem
+Revise daemon code
