+++ daemon/src/imgproc/mod.rs
@@ -51,14 +51,14 @@ impl Imgproc {
                     }
                     let dimensions = wallpapers[0].get_dimensions();
 
-                    if img.len() == dimensions.0 as usize * dimensions.1 as usize * 4 {
+                    if img.len() == dimensions.0 as usize * dimensions.1 as usize * 3 {
                         animations::Transition::new(wallpapers, dimensions, transition, pool)
                             .execute(&img);
                     } else {
                         error!(
                             "image is of wrong size! Image len: {}, expected size: {}",
                             img.len(),
-                            dimensions.0 as usize * dimensions.1 as usize * 4
+                            dimensions.0 as usize * dimensions.1 as usize * 3
                         );
                     }
                 })
@@ -104,7 +104,10 @@ impl Imgproc {
                     for wallpaper in wallpapers.iter_mut() {
                         let mut pool = wallpaper.lock_pool_to_get_canvas(&pool);
                         let canvas = wallpaper.get_canvas(&mut pool);
-                        frame.ready(canvas.len()).unpack(canvas);
+                        if !frame.unpack(canvas) {
+                            error!("failed to unpack frame, canvas has the wrong size");
+                            return;
+                        }
                         wallpaper.draw(&mut pool);
                     }
