+++ utils/benches/compression.rs
@@ -1,8 +1,8 @@
-use criterion::{black_box, criterion_group, criterion_main, BatchSize, Criterion};
-use utils::comp_decomp::{BitPack, ReadiedPack};
+use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use utils::comp_decomp::BitPack;
 
 fn generate_data() -> (Box<[u8]>, Box<[u8]>) {
-    let v1 = vec![120; 1920 * 1080 * 4];
+    let v1 = vec![120; 1920 * 1080 * 3];
     let mut v2 = v1.clone();
 
     const REGIONS: usize = 2000;
@@ -27,52 +27,34 @@ fn generate_data() -> (Box<[u8]>, Box<[u8]>) {
     (v1.into_boxed_slice(), v2.into_boxed_slice())
 }
 
+fn buf_from(slice: &[u8]) -> Vec<u8> {
+    let mut v = Vec::new();
+    for pix in slice.chunks_exact(3) {
+        v.extend_from_slice(pix);
+        v.push(255);
+    }
+    v
+}
+
 pub fn compression_and_decompression(c: &mut Criterion) {
     let (prev, cur) = generate_data();
 
     let mut comp = c.benchmark_group("compression");
     comp.bench_function("Full", |b| {
-        b.iter_batched(
-            || prev.clone(),
-            |mut prev| {
-                black_box(BitPack::pack(&mut prev, &cur).ok());
-            },
-            BatchSize::SmallInput,
-        )
-    });
-    comp.bench_function("Partial", |b| {
-        b.iter_batched(
-            || prev.clone(),
-            |mut prev| {
-                black_box(ReadiedPack::new(&mut prev, &cur, |_, _, _| {}));
-            },
-            BatchSize::SmallInput,
-        )
+        b.iter(|| {
+            black_box(BitPack::pack(&prev, &cur).ok());
+        })
     });
     comp.finish();
 
     let mut decomp = c.benchmark_group("decompression");
-    let mut p = prev.clone();
-    let bitpack = BitPack::pack(&mut p, &cur).unwrap();
-    let ready = bitpack.ready(prev.len());
+    let bitpack = BitPack::pack(&prev, &cur).unwrap();
+    let mut canvas = buf_from(&prev);
 
     decomp.bench_function("Full", |b| {
-        b.iter_batched(
-            || prev.clone(),
-            |mut prev| {
-                black_box(bitpack.ready(prev.len()).unpack(&mut prev));
-            },
-            BatchSize::SmallInput,
-        )
-    });
-    decomp.bench_function("Partial", |b| {
-        b.iter_batched(
-            || prev.clone(),
-            |mut prev| {
-                black_box(ready.unpack(&mut prev));
-            },
-            BatchSize::SmallInput,
-        )
+        b.iter(|| {
+            black_box(bitpack.unpack(&mut canvas));
+        })
     });
 
     decomp.finish();
